hyper-v: Converts all os-win exceptions
When os-win was first introduced, a decorator which converts all os-win
exceptions to virt inspector exceptions has been added to the
HyperVInspector.

However, the decorator works as intended for methods that return values,
and not for the ones that yield (they return a generator).

This patch makes sure that exceptions are converted properly for yielding
methods as well.

Closes-Bug #1751088

Change-Id: I7d09e1860c6940758f0d0965fedfe4dd285e0cae